+++ src/net/sf/persism/Reader.java
@@ -1,26 +1,21 @@
 package net.sf.persism;
 
-import net.sf.persism.annotations.NotTable;
-
-import java.beans.ConstructorProperties;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringWriter;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Parameter;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.sql.*;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 final class Reader {
 
     private static final Log log = Log.getLogger(Reader.class);
-    private static final Log blog = Log.getLogger("net.sf.persism.Benchmarks");
+//    private static final Log blog = Log.getLogger("net.sf.persism.Benchmarks");
 
     private Connection connection;
     private MetaData metaData;
@@ -35,6 +30,9 @@ final class Reader {
     }
 
     <T> T readObject(T object, Map<String, PropertyInfo> properties, ResultSet rs) throws SQLException, IOException {
+
+        long now = System.nanoTime();
+
         Class<?> objectClass = object.getClass();
 
         // We should never call this method with a primitive type.
@@ -69,6 +67,10 @@ final class Reader {
             // Save this object initial state to later detect changed properties
             pojo.saveReadState();
         }
+
+        // blog.debug("OBJECT: time to read all columns: %s", (System.nanoTime() - now));
+        //System.out.println("time to readObject: " + (System.nanoTime() - now));
+
         return (T) object;
     }
 
@@ -78,7 +80,7 @@ final class Reader {
         now = System.nanoTime();
 
         ResultSetMetaData rsmd = rs.getMetaData();
-        List<Object> constructorParams = new ArrayList<>(12);
+        List<Object> constructorParams = new ArrayList<>(recordInfo.propertyInfoByConstructorOrder.keySet().size());
 
         for (String col : recordInfo.propertyInfoByConstructorOrder.keySet()) {
             Class<?> returnType = recordInfo.propertyInfoByConstructorOrder.get(col).field.getType();
@@ -93,12 +95,16 @@ final class Reader {
             constructorParams.add(value);
         }
 
+        // blog.debug("RECORD: time to read all columns: %s", (System.nanoTime() - now));
+
+        //      now = System.nanoTime();
 
         try {
             //noinspection unchecked
-            return (T) recordInfo.constructor.newInstance(constructorParams.toArray());
+            return recordInfo.constructor.newInstance(constructorParams.toArray());
         } finally {
-            blog.debug("time to get readRecord: %s", (System.nanoTime() - now));
+            // blog.debug("time to get readRecord: %s", (System.nanoTime() - now));
+            // System.out.println("time to readRecord: " + (System.nanoTime() - now));
         }
     }
 
@@ -249,8 +255,8 @@ final class Reader {
             value = converter.convert(value, returnType, columnName);
         }
 
-        blog.debug("time to readColumn: %s", (System.nanoTime() - now));
-        return  value;
+        //blog.debug("time to readColumn: %s", (System.nanoTime() - now));
+        return value;
     }