+++ src/net/sf/persism/annotations/NotColumn.java
@@ -4,7 +4,7 @@ import java.lang.annotation.*;
 
 /**
  * Indicates that the property is not mapped to a column. This annotation can appear on the class field, getter or setter.
- * Persism ignores read-only properties so this annotation is only required where you have a read/write property which does not map to any column.
+ * It works the same if you use the transient keyword on a field.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(value = {ElementType.METHOD, ElementType.FIELD})