+++ test/net/sf/persism/dao/so/LinkType.java
@@ -1,0 +1,7 @@
+package net.sf.persism.dao.so;
+
+public record LinkType(Integer id, String type) {
+
+    public LinkType {
+    }
+}