+++ test/net/sf/persism/dao/so/VoteType.java
@@ -1,0 +1,50 @@
+package net.sf.persism.dao.so;
+
+public class VoteType {
+    private Integer id;
+    private String name;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        VoteType voteTypes = (VoteType) o;
+
+        if (id != null ? !id.equals(voteTypes.id) : voteTypes.id != null) {
+            return false;
+        }
+        if (name != null ? !name.equals(voteTypes.name) : voteTypes.name != null) {
+            return false;
+        }
+
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        int result = id != null ? id.hashCode() : 0;
+        result = 31 * result + (name != null ? name.hashCode() : 0);
+        return result;
+    }
+}